feat: NeomClaw v0.1.0 — multi-provider AI coding assistant
Complete migration from OpenClaude to Flutter (283 files, ~179K LOC):

- Multi-provider: Gemini (default), Qwen, OpenAI, DeepSeek, Anthropic, Ollama
- Agentic tool loop: bash, file read/write/edit, grep, glob with streaming
- Ollama local model manager: auto-discover, download, test, activate models
- Image/file/PDF attachments with preview in chat input
- MCP (Model Context Protocol) client panel with stdio/SSE transport
- Settings with provider config, connection testing, diagnostics
- Onboarding flow with provider selection
- Cross-platform: macOS, Linux, Windows, Web, iOS, Android
- Dark/light Material 3 theme with design system
- Rebranded: CLAUDE.md → NEOMCLAW.md, .claude/ → .neomclaw/
- Removed neom_core path dependency — ready for pub.dev
- Added LICENSE (MIT), CHANGELOG.md, pub.dev metadata

+++ lib/ui/vim/vim_mode.dart
@@ -1,5 +1,5 @@
 // vim_mode.dart — Vim emulation layer for flutter_claw
-// Port of openclaude/src/vim/ (~1.5K TS LOC) to pure Dart + minimal Flutter.
+// Port of neom_claw/src/vim/ (~1.5K TS LOC) to pure Dart + minimal Flutter.
 
 import 'dart:math';
 
@@ -685,7 +685,7 @@ class VimKeyHandler {
       case CharMotion m:
         final newCol = (state.cursor.column + m.delta * count)
             .clamp(0, max(0, _currentLineLength() - 1));
-        state.cursor = state.cursor.copyWith(column: newCol);
+        state.cursor = state.cursor.copyWith(column: newCol.toInt());
       case LineMotion m:
         final newLine =
             (state.cursor.line + m.delta * count).clamp(0, _lines.length - 1);
